requirements.txt
python-dotenv==1.0.1
pandas~=2.2.2
requests~=2.32.3
tabulate~=0.9.0
peewee~=3.17.5